+++ client/package.json
@@ -11,7 +11,8 @@
     "react": "^16.6.3",
     "react-dom": "^16.6.3",
     "react-router-dom": "^4.3.1",
-    "react-scripts": "^2.1.1"
+    "react-scripts": "^2.1.1",
+    "showdown": "^1.9.0"
   },
   "scripts": {
     "start": "react-scripts start",
