+++ client/public/chat.html
@@ -1,12 +1,14 @@
 <html>
     <head>
         <title>Botkit Anywhere</title>
+        
         <script src="https://cdnjs.cloudflare.com/ajax/libs/handlebars.js/4.0.10/handlebars.min.js"></script>
         <script src="https://cdn.rawgit.com/showdownjs/showdown/1.7.4/dist/showdown.min.js"></script>
-        <link rel="stylesheet" href="css/styles.css" />
+        <link rel="stylesheet" href="./css/styles.css" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
     </head>
     <body >
+      
       <div class="wrapper">
 
         <div id="message_window">
@@ -58,6 +60,7 @@
             </footer>
         </div>
         </div>
-        <script src="client.js"></script>
+        <script src="./js/client.js"></script>
+        <script src="./js/embed.js"></script>
     </body>
 </html>
