+++ client/public/index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
     <meta name="theme-color" content="#000000">
 
-    <link rel="stylesheet" href="./css/embed.css" />
+    <link rel="stylesheet" href="%PUBLIC_URL%/embed.css" />
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json">
     <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
@@ -14,14 +14,14 @@
   </head>
   <body>
 
-    <script src="./embed.js"></script>
-
-    <script>
-        var options = {};
-        Botkit.boot(options);
-    </script>
-
     <div id="root"></div>
+
+    <script src="%PUBLIC_URL%/embed.js"></script>   
+
+        <script>
+            var options = {};
+            Botkit.boot(options);
+        </script>
 
   </body>
 </html>
