+++ client/src/index.js
@@ -5,4 +5,5 @@ import App from "./App";
 import registerServiceWorker from "./registerServiceWorker";
 
 ReactDOM.render(<App />, document.getElementById("root"));
+//ReactDOM.render(<App />, document.getElementById("wrapper"));
 registerServiceWorker();
