+++ client/src/sass/_chat.scss
@@ -1,3 +1,4 @@
+$botkit_purple: #a795ef;
 $action_color: orange;
 $background: #FFFFFF;
 $bot_message_background: #F0F0F0;
